pre-research/requirements.txt
# Core dependencies
requests>=2.31.0
click>=8.1.0
rich>=13.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Optional dependencies
pytest-watch>=4.2.0  # For watch mode during development